eliminate gcc warning

+++ src/proto/socks5address.cpp
@@ -46,7 +46,7 @@ bool SOCKS5Address::parse(const string &data, size_t &address_len) {
                 // invalid domain len
                 break;
             }
-            if (data.length() > 1 + domain_len + 2) {
+            if (data.length() > (unsigned int)(1 + domain_len + 2)) {
                 address = data.substr(2, domain_len);
                 port = (uint8_t(data[domain_len + 2]) << 8) | uint8_t(data[domain_len + 3]);
                 address_len =  1 + 1 + domain_len + 2;
